More control stuff, every int is now a long

+++ src/benchmark/Benchmark.java
@@ -2,8 +2,9 @@ package benchmark;
 
 public interface Benchmark {
 
-    int runTest();
-    void setScore(int score);
+    long runTest();
+
+    void setScore(long score);
 
     String getName();
